+++ internal/git/phantom_conflict.go
@@ -4,6 +4,7 @@ import (
 	"reflect"
 
 	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
 )
 
 // describes a file within an unresolved merge conflict that experiences a phantom merge conflict
